refactor: extract ensureGistPersistence helper (single-source for CLI + MCP)
Closes #1000.

The "peek at state file, if persistence === 'gist' then call
getStateManagerAsync" pattern was duplicated between cli.ts:62-77 and
mcp-server/src/tools.ts:36-56. Two separate copies of ~15 lines of
filesystem + JSON parsing logic at the boundary between startup and
core state management.

Extract ensureGistPersistence(token) to packages/core/src/core/state.ts
(colocated with getStateManagerAsync), re-export via core/index.ts, and
collapse both call sites to a single await.

- cli.ts: 16 lines -> 4 lines
- mcp-server tools.ts: 22 lines -> 9 lines (still wraps with gistInitDone
  one-shot guard since MCP tool calls can be repeated in a long-lived
  process, unlike the CLI's single-invocation lifecycle)

MCP test mocks in tools-execution.test.ts, prompts.test.ts, and
resources.test.ts now include getGitHubTokenAsync + ensureGistPersistence
stubs so wrapTool's ensureGistInit call resolves cleanly.
